Create Metagenomic shotgun QC
R file from 1 March 2023, last used to make the plots for the manuscript. Have removed the code for bar plot and heat map as it is the same as previous. Just changed the variables. This example is with Genus but if you change to species it will do the pipeline from species level.